# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Mycorrhizal CRM API
  description: >
    REST API for Mycorrhizal's contact model. As of WP-71 (P2,
    docs/fork-plan/50-integration-and-rebrand.md), the contact endpoints speak
    the neutral RFC 9553-shaped model (contactmodel.Record/Card) rather than
    the legacy flat Contact DTO. GET /contacts (list) returns a slim
    ContactSummary per item; GET/POST/PUT /contacts/{id} exchange the full
    Card/CRM shape (ContactRecordResponse / ContactRecordInput) — a
    symmetric read/write contract.


    T8 (docs/fork-plan/tickets/16-T8-openapi.md) extended this spec to cover
    the whole REST surface: every route registered by
    backend/routes/routes.go under /api/v1 plus the unversioned /health. The
    drift test (backend/openapi_test.go) enumerates the live Gin router and
    fails on any registered route with no spec entry, so this file cannot go
    stale again.

    Coverage notes:

    - **Auth:** every path under /api/v1 except the public auth group
      (register, login, logout, check-password-strength, password-reset,
      auth/oidc/config) requires a session cookie (`auth_token`, set by login
      and OIDC callback) or an `Authorization: Bearer <token>` header. API
      tokens (mycorrhizal_…) are accepted on the general REST path only when
      their scope is `full`; a `carddav`-scoped token is rejected with 403.
      API tokens cannot access the admin paths at all.
...
      `/auth/oidc/config` is always present.
...
      them.
...
servers:
...
security:
  - sessionCookie: []
  - bearerAuth: []
paths:
...
  /circles:
    get:
      summary: List circles
      operationId: listCircles
      description: >
        Cursor-paginated (T17). Circles are a hard-delete collection — the
        sync.mode is full_resync: a client re-pulls the collection rather than
        using ?since= (circles have no tombstone).
      parameters:
        - name: cursor
          in: query
          description: Opaque resume token from the previous page's next_cursor.
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, default: 25 }
        - name: order
          in: query
          description: Direction for the (updated_at, id) cursor order.
          schema: { type: string, enum: [asc, desc], default: desc }
        - name: include_members
          in: query
          description: When true, also returns a members array (all circle_members join rows for the page's circles).
          schema: { type: boolean, default: false }
      responses:
        "401":
          description: Not authenticated (missing or invalid session cookie or full-scope bearer token).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "200":
          description: A page of circles.
          content:
            application/json:
              schema:
                type: object
                properties:
                  circles:
                    type: array
                    items:
                      $ref: "#/components/schemas/Circle"
                  total: { type: integer }
                  next_cursor:
                    type: string
                    description: Resume token for the next page; empty when there are no more rows.
                  limit: { type: integer }
                  sync:
                    $ref: "#/components/schemas/SyncInfo"
                  members:
                    type: array
                    items:
                      $ref: "#/components/schemas/CircleMember"
    post:
      summary: Create a circle
      operationId: createCircle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CircleInput"
      responses:
        "200":
          description: Circle created (200, not 201).
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: "Circle created successfully" }
                  circle:
                    $ref: "#/components/schemas/Circle"
        "400":
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /circles/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Get a circle with its members
      operationId: getCircle
      responses:
        "200":
          description: The circle and its membership join rows.
          content:
            application/json:
              schema:
                type: object
                properties:
                  circle:
                    $ref: "#/components/schemas/Circle"
                  members:
                    type: array
                    items:
                      $ref: "#/components/schemas/CircleMember"
        "404":
          description: Circle not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Update a circle
      operationId: updateCircle
      description: Only name is editable; membership lives in the member endpoints.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CircleInput"
      responses:
        "200":
          description: The updated circle (raw, not wrapped).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Circle"
        "400":
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Circle not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete a circle (hard delete; members cascade)
      operationId: deleteCircle
      responses:
        "200":
          description: Deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Circle not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /circles/{id}/members:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      summary: Add a member to a circle
      operationId: addCircleMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CircleMemberInput"
      responses:
        "200":
          description: Member added.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: "Member added" }
                  member:
                    $ref: "#/components/schemas/CircleMember"
        "400":
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Circle or contact not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: The contact is already a member of this circle.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /circles/{id}/members/{vcard_uid}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: vcard_uid
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      summary: Remove a member from a circle
      operationId: removeCircleMember
      responses:
        "200":
          description: Member removed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Circle or membership not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # --- Households ---
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: auth_token
      description: The httpOnly session cookie set by POST /login and the OIDC callback.
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        A JWT or a `mycorrhizal_…` API token. On the general REST path a
        carddav-scoped token is rejected with 403 (only `full`-scope tokens
        work); API tokens are rejected on every /admin path entirely.
  schemas:
...
    Error:
      type: object
      description: The standard error envelope produced by apperrors.AbortWithError.
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: object
              additionalProperties: true
        request_id: { type: string }
        timestamp: { type: string, format: date-time }
    SyncInfo:
      type: object
      description: >
        T17 sync-mode contract carried by every cursor-paginated list
        response. `mode` is how THIS collection syncs; the two arrays are the
        static map of every syncable collection, so a client can plan a full
        sync without hardcoding endpoint names. Incremental collections are
        soft-deleted, support ?since= change feeds, and drop `total`;
        full_resync collections are hard-deleted, are re-pulled wholesale, and
        keep `total`.
      properties:
        mode: { type: string, enum: [incremental, full_resync] }
        incremental:
          type: array
          items: { type: string }
          description: "Collections that sync incrementally via ?since=: contacts, notes, activities, life_events, preferences."
        full_resync:
          type: array
          items: { type: string }
          description: "Collections that require a full re-pull: circles, households, tags, relationship_edges, field_definitions."
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }
      additionalProperties: true
...
    Circle:
      type: object
      properties:
        id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        name: { type: string }
    CircleInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
    CircleMember:
      type: object
      description: A circle membership join row.
      properties:
        id: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        circle_id: { type: string, format: uuid }
        member_vcard_uid: { type: string, format: uuid }
    CircleMemberInput:
      type: object
      required: [member_vcard_uid]
      properties:
        member_vcard_uid: { type: string, format: uuid }
...